Responsive list added
