+++ src/Components/News/index.js
@@ -1,0 +1,11 @@
+import Title from '../Titles'
+import NewsList from './NewsList';
+
+export const NewsScreen = () => {     
+    return (
+        <>
+            <Title children="Novedades"/>
+            <NewsList  />        
+        </>
+    )
+}
